merge: merge cleanup
